renamed some files, because I have strong opinions disabled nyan cat, sorry nyan cat
