+++ app.js
@@ -13,7 +13,7 @@ app.set('view engine', 'jade');
 
 // uncomment after placing your favicon in /public
 //app.use(favicon(path.join(__dirname, 'public', 'favicon.ico')));
-// app.use(logger('dev'));
+app.use(logger('dev'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
@@ -31,7 +31,7 @@ app.use('/all/registered/:page', zomgRoutes);
 var users = require('./routes/users');
 app.use('/users', users);
 
-var movies = require('./routes/movieRoutes');
+var movies = require('./routes/movie_routes');
 app.use('/movies', movies);
